refactor(ClassFileName): Use upstream Squiz ClassFileName sniff directly
