--- tests/Drupal/Classes/drupal7/class_file_name_test.module
@@ -1,5 +1,0 @@
-<?php
-
-class OkName {
-
-}
