--- tests/Drupal/Classes/drupal7/drupal7.info
@@ -1,1 +1,0 @@
-core = 7.x
