--- tests/Drupal/Classes/drupal8/ClassFileNameUnitTest.php
@@ -1,5 +1,0 @@
-<?php
-
-class WrongName {
-
-}
