--- tests/Drupal/Classes/drupal8/drupal8.behat.inc
@@ -1,10 +1,0 @@
-<?php
-
-use Drupal\DrupalExtension\Context\DrupalSubContextBase;
-
-/**
- * Step definitions for testing the functionality of the Drupal8 module.
- */
-class Drupal8SubContext extends DrupalSubContextBase {
-
-}
