--- tests/Drupal/Classes/drupal8/drupal8.info.yml
@@ -1,2 +1,0 @@
-core: 8.x
-name: Test
