--- tests/Drupal/Classes/drupal8/markdownFile.md
@@ -1,11 +1,0 @@
-this is not a php file
-
-```php
-<?php
-
-/**
- * Class.
- */
-class Test {
-}
-```
