+++ tests/Drupal/good/GoodAlignment.php
@@ -1,0 +1,24 @@
+<?php
+
+/**
+ * Nice alignment is allowed for assignments.
+ */
+class GoodAlignment {
+  const INFO      = 0;
+  const WARNING   = 1;
+  const ERROR     = 2;
+  const EMERGENCY = 3;
+
+  /**
+   * Empty method implementation is allowed.
+   */
+  public function emptyMethod() {}
+
+  /**
+   * Protected functions are allowed.
+   */
+  protected function protectedTest() {
+
+  }
+
+}
