+++ tests/Drupal/good/GoodClassOperator.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * Test class.
+ */
+class GoodClassOperator {
+
+  /**
+   * Seen IDs.
+   *
+   * @var array
+   */
+  protected static $seenIds;
+
+  /**
+   * Test method.
+   */
+  public function test() {
+    $id = $id . '--' . ++static::$seenIds[$id];
+    return $id;
+  }
+
+}
