+++ tests/Drupal/good/GoodCoversTag.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * @coversDefaultClass \Drupal\rules\Plugin\Condition\ListContains
+ * @group rules_conditions
+ */
+class GoodCoversTag extends RulesIntegrationTestBase {
+
+  /**
+   * @covers ::foo
+   */
+  public function testFoo() {
+
+  }
+
+}
