+++ tests/Drupal/good/GoodEnum.php
@@ -1,0 +1,9 @@
+<?php
+
+/**
+ * Enums must have a comment block.
+ */
+enum GoodEnum: int {
+  case One = 1;
+  case Two = 2;
+}
