+++ tests/Drupal/good/GoodEnumNameIgnore.php
@@ -1,0 +1,10 @@
+<?php
+
+/**
+ * Doc block is here and an ignore directive is ok.
+ */
+// phpcs:ignore Drupal.NamingConventions.ValidClassName,Squiz.Classes.ClassFileName.NoMatch
+enum PUROSELY_WRONG_BUT_OK: int {
+  case One = 1;
+  case Two = 2;
+}
