+++ tests/Drupal/good/GoodException.php
@@ -1,0 +1,6 @@
+<?php
+
+/**
+ * It is allowed to have the closing "}" on the same line if the class is empty.
+ */
+class GoodException extends Exception {}
