+++ tests/Drupal/good/GoodInterface.php
@@ -1,0 +1,6 @@
+<?php
+
+/**
+ * Interfaces must have a comment block.
+ */
+interface GoodInterface {}
