+++ tests/Drupal/good/GoodLongAnnotation.php
@@ -1,0 +1,53 @@
+<?php
+
+/**
+ * Example annotation that exceeds 80 characters several times, but is valid.
+ *
+ * @ConfigEntityType(
+ *   id = "rules_reaction_rule",
+ *   label = @Translation("Reaction Rule"),
+ *   handlers = {
+ *     "storage" = "Drupal\rules\Entity\ReactionRuleStorage",
+ *     "list_builder" = "Drupal\rules\Entity\Controller\RulesReactionListBuilder",
+ *     "form" = {
+ *        "add" = "\Drupal\rules\Entity\ReactionRuleAddForm",
+ *        "edit" = "\Drupal\rules\Entity\ReactionRuleEditForm",
+ *        "delete" = "\Drupal\Core\Entity\EntityDeleteForm"
+ *      }
+ *   },
+ *   cardinality = \Drupal\webform\WebformHandlerInterface::CARDINALITY_UNLIMITED,
+ *   admin_permission = "administer rules",
+ *   config_prefix = "reaction",
+ *   entity_keys = {
+ *     "id" = "id",
+ *     "label" = "label",
+ *     "status" = "status"
+ *   },
+ *   config_export = {
+ *     "id",
+ *     "label",
+ *     "event",
+ *     "module",
+ *     "description",
+ *     "tag",
+ *     "core",
+ *     "expression_id",
+ *     "configuration",
+ *   },
+ *   links = {
+ *     "collection" = "/admin/config/workflow/rules",
+ *     "edit-form" = "/admin/config/workflow/rules/reactions/edit/{rules_reaction_rule}",
+ *     "delete-form" = "/admin/config/workflow/rules/reactions/delete/{rules_reaction_rule}"
+ *   }
+ * )
+ */
+class GoodLongAnnotation extends ConfigEntityBase {
+
+  /**
+   * Config entities are allowed to have property names with underscores.
+   *
+   * @var string
+   */
+  protected $expression_id;
+
+}
