+++ tests/Drupal/good/good.php
@@ -579,108 +579,6 @@ $var = foo(
 );
 
 /**
- * Class declaration.
- *
- * Classes can have a multiline comment.
- */
-class Bar {
-
-  /**
-   * Private properties have no prefix.
-   *
-   * @var int
-   */
-  private $secret = 1;
-
-  /**
-   * Protected properties also don't have a prefix.
-   *
-   * @var int
-   */
-  protected $foo = 1;
-
-  /**
-   * Longer properties use camelCase naming.
-   *
-   * @var int
-   */
-  public $barProperty = 1;
-
-  /**
-   * Using property types is allowed.
-   *
-   * @var \Foo\Bar
-   */
-  public ?Bar $bar;
-
-  /**
-   * A typed property may omit @var.
-   */
-  public Bar $baz;
-
-  /**
-   * Public static variables use camelCase, too.
-   *
-   * @var string
-   */
-  public static $basePath = NULL;
-
-  /**
-   * {@inheritdoc}
-   */
-  protected $modules = ['node', 'user'];
-
-  /**
-   * {@inheritDoc}
-   */
-  protected $allowedModules = ['node', 'user'];
-
-  /**
-   * Enter description here ...
-   */
-  public function foo() {
-
-  }
-
-  /**
-   * Enter description here ...
-   */
-  protected function barMethod() {
-
-  }
-
-  /**
-   * Test the ++ and -- operator.
-   */
-  public function incDecTest() {
-    $this->foo++;
-    $this->foo--;
-    --$this->foo;
-    ++$this->foo;
-  }
-
-  /**
-   * It is allowed to leave out param docs on methods.
-   */
-  public function noParamDocs($a, $b) {
-
-  }
-
-  /**
-   * Param comments with references are found correctly.
-   *
-   * @param string $a
-   *   Parameter one.
-   * @param array $b
-   *   Parameter two.
-   */
-  public function test($a, array &$b) {
-
-  }
-
-}
-
-/**
  * Enter description here ...
  */
 function _private_foo() {
@@ -705,34 +603,6 @@ $i . "test";
 $i . 'test';
 $i . $i;
 $i . NULL;
-
-/**
- * It is allowed to have the closing "}" on the same line if the class is empty.
- */
-class MyException extends Exception {}
-
-/**
- * Nice alignment is allowed for assignments.
- */
-class MyExampleLog {
-  const INFO      = 0;
-  const WARNING   = 1;
-  const ERROR     = 2;
-  const EMERGENCY = 3;
-
-  /**
-   * Empty method implementation is allowed.
-   */
-  public function emptyMethod() {}
-
-  /**
-   * Protected functions are allowed.
-   */
-  protected function protectedTest() {
-
-  }
-
-}
 
 /**
  * Nice allignment in functions.
@@ -982,11 +852,6 @@ preg_filter('@.+@i', 'replacement', 'subject');
 preg_replace('@.+@i', 'replacement', 'subject');
 preg_match("/test(\d+)/is", 'subject');
 
-/**
- * Interfaces must have a comment block.
- */
-interface MyWellNamedInterface {}
-
 // Correctly formed try/catch block.
 try {
   do_something();
@@ -997,134 +862,7 @@ catch (Exception $e) {
 
 $result = $x ?: FALSE;
 
-/**
- * All classes need to have a docblock.
- */
-class Foo implements FooInterface {
-
-  /**
-   * {@inheritdoc}
-   */
-  public function test() {
-    /** @var \Drupal\node\NodeInterface $node */
-    $node = $this->entity;
-    /** @var \Drupal\node\NodeInterface[] $nodes */
-    $nodes = foo();
-    /** @var \Drupal\node\NodeInterface|\PHPUnit_Framework_MockObject_MockObject $node_mock */
-    $node_mock = mock_node();
-    /** @var \Drupal\SomeInterface4You $thing */
-    $thing = thing();
-    /** @var \Drupal\SomeInterface4You $test2 */
-    $test2 = test2();
-    return $node;
-  }
-
-  /**
-   * {@inheritdoc}
-   *
-   * Some additional documentation here.
-   */
-  public function test2() {}
-
-  /**
-   * Return docs are allowed to use $this.
-   *
-   * @return $this
-   *   This object for chaining method calls.
-   */
-  public function test3() {
-    return $this;
-  }
-
-  /**
-   * Returns the string representatuion of this object.
-   */
-  public function __toString() {
-    return 'foo';
-  }
-
-  /**
-   * Omitting the comment when returning $this is allowed.
-   *
-   * @return $this
-   */
-  public function test4() {
-    return $this;
-  }
-
-  /**
-   * Omitting the comment when returning static is allowed.
-   *
-   * @return static
-   */
-  public function test41() {
-    return new static();
-  }
-
-  /**
-   * Loads multiple string objects.
-   *
-   * @param array $conditions
-   *   Any of the conditions used by dbStringSelect().
-   * @param array $options
-   *   Any of the options used by dbStringSelect().
-   * @param string $class
-   *   Class name to use for fetching returned objects.
-   *
-   * @return \Drupal\locale\StringInterface[]
-   *   Array of objects of the class requested.
-   */
-  protected function dbStringLoad(array $conditions, array $options, $class) {
-    $strings = [];
-    $result = $this->dbStringSelect($conditions, $options)->execute();
-    foreach ($result as $item) {
-      /** @var \Drupal\locale\StringInterface $string */
-      $string = new $class($item);
-      $string->setStorage($this);
-      $strings[] = $string;
-    }
-    return $strings;
-  }
-
-  /**
-   * Short array syntax is allowed.
-   */
-  public function getConfiguration() {
-    return [
-      'id' => $this->getPluginId(),
-    ] + $this->configuration;
-  }
-
-  /**
-   * Not documenting a "throws" tag is allowed.
-   *
-   * This should not fail for errors with underscores in names as well.
-   * The second version of this test with error name with underscores
-   * is added below.
-   *
-   * @throws Exception
-   */
-  public function test6() {
-    throw new Exception();
-  }
-
-  /**
-   * Repeat of above test with error name with underscores.
-   *
-   * @throws \Twig_Error_Syntax
-   */
-  public function test7() {
-    throw new Exception();
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  public function test8() {}
-
-}
-
-t('Some long mulit-line 
+t('Some long mulit-line
   text is weird, but allowed.');
 
 // Anonymous functions should not throw indentation errors here.
@@ -1199,32 +937,7 @@ function test7($x) {
 
 }
 
-/**
- * @coversDefaultClass \Drupal\rules\Plugin\Condition\ListContains
- * @group rules_conditions
- */
-class ListContainsTest extends RulesIntegrationTestBase {}
-
 $x = 'Some markup text with allowed HTML5 <br> tag';
-
-/**
- * Provides a 'Delete any path alias' action.
- *
- * @todo Add access callback information from Drupal 7.
- * @todo Add group information from Drupal 7.
- *
- * @Action(
- *   id = "rules_path_alias_delete",
- *   label = @Translation("Delete any path alias"),
- *   context = {
- *     "alias" = @ContextDefinition("string",
- *       label = @Translation("Existing system path alias"),
- *       description = @Translation("Specifies the existing path alias you wish to delete, for example 'about/team'. Use a relative path and do not add a trailing slash.")
- *     )
- *   }
- * )
- */
-class AliasDelete extends RulesActionBase implements ContainerFactoryPluginInterface {}
 
 /**
  * Some comment with exclamation mark!
@@ -1307,93 +1020,6 @@ function test13($a) {
  *   Example parameter.
  */
 function test14(AliasedExampleClass $a) {
-
-}
-
-/**
- * Example annotation that exceeds 80 characters several times, but is valid.
- *
- * @ConfigEntityType(
- *   id = "rules_reaction_rule",
- *   label = @Translation("Reaction Rule"),
- *   handlers = {
- *     "storage" = "Drupal\rules\Entity\ReactionRuleStorage",
- *     "list_builder" = "Drupal\rules\Entity\Controller\RulesReactionListBuilder",
- *     "form" = {
- *        "add" = "\Drupal\rules\Entity\ReactionRuleAddForm",
- *        "edit" = "\Drupal\rules\Entity\ReactionRuleEditForm",
- *        "delete" = "\Drupal\Core\Entity\EntityDeleteForm"
- *      }
- *   },
- *   admin_permission = "administer rules",
- *   config_prefix = "reaction",
- *   entity_keys = {
- *     "id" = "id",
- *     "label" = "label",
- *     "status" = "status"
- *   },
- *   config_export = {
- *     "id",
- *     "label",
- *     "event",
- *     "module",
- *     "description",
- *     "tag",
- *     "core",
- *     "expression_id",
- *     "configuration",
- *   },
- *   links = {
- *     "collection" = "/admin/config/workflow/rules",
- *     "edit-form" = "/admin/config/workflow/rules/reactions/edit/{rules_reaction_rule}",
- *     "delete-form" = "/admin/config/workflow/rules/reactions/delete/{rules_reaction_rule}"
- *   }
- * )
- */
-class ReactionRule extends ConfigEntityBase {
-
-  /**
-   * Config entities are allowed to have property names with underscores.
-   *
-   * @var string
-   */
-  protected $expression_id;
-
-}
-
-/**
- * Underscores are allowed in properties of config entity classes.
- */
-class NodeType extends ConfigEntityBundleBase {
-
-  /**
-   * Default value of the 'Create new revision' checkbox of this node type.
-   *
-   * @var bool
-   */
-  protected $new_revision = TRUE;
-
-}
-
-/**
- * Test class.
- */
-class OperatorTest {
-
-  /**
-   * Seen IDs.
-   *
-   * @var array
-   */
-  protected static $seenIds;
-
-  /**
-   * Test method.
-   */
-  public function test() {
-    $id = $id . '--' . ++static::$seenIds[$id];
-    return $id;
-  }
 
 }
 
@@ -1539,20 +1165,6 @@ function mymodule_very_long_module_name_neverending_hook_name_that_is_a_long_cat
 }
 
 /**
- * @coversDefaultClass \Drupal\mymodule\Something
- */
-class MyTest extends UnitTestBase {
-
-  /**
-   * @covers ::foo
-   */
-  public function testFoo() {
-
-  }
-
-}
-
-/**
  * Numbers in types are allowed.
  *
  * @param \Drupal\x123_module\MyInterface $a
@@ -1582,285 +1194,12 @@ function test24(Element ...$element) {
 
 }
 
-/**
- * Test class.
- */
-class Test2 {
-
-  /**
-   * Using PHP 7 return type hints is fine.
-   *
-   * @return ValidatorInterface[]
-   *   The validators.
-   */
-  public function getValidators(): array {
-    return [
-      new PublishedNodesValidator(),
-      new MinimumNodesValidator($this->nrOfArticles),
-      new AccessibleOnCurrentDomainValidator($this->sectionService),
-    ];
-  }
-
-}
-
-/**
- * Another test.
- */
-class Test3 {
-
-  /**
-   * Parameters described by reference are OK.
-   *
-   * @param array &$form
-   *   The form array.
-   * @param Drupal\Core\Form\FormStateInterface $form_state
-   *   The form state.
-   *
-   * @return array
-   *   A renderable array.
-   */
-  public function removeQueueItem(array &$form, FormStateInterface $form_state) {
-    $trigger = $form_state->getTriggeringElement();
-    $i = $trigger['#parents'][1];
-
-    $queues = $form_state->getValue('watch_queues', []);
-    $queues[$i]['to_remove'] = 1;
-    $form_state->setValue('watch_queues', $queues);
-    $this->rebuild($form_state, $form);
-
-    drupal_set_message($this->t('Item will be removed permanently when configuration is saved.'));
-    return StatusMessages::renderMessages(NULL);
-  }
-
-  /**
-   * Parameters described by reference are OK.
-   *
-   * @param Drupal\Core\Form\FormStateInterface $form_state
-   *   The form state.
-   * @param array &$old_form
-   *   The old form build.
-   *
-   * @return array
-   *   The newly built form.
-   */
-  protected function rebuild(FormStateInterface $form_state, array &$old_form) {
-    $form_state->setRebuild();
-    $form = $this->formBuilder
-      ->rebuildForm($this->getFormId(), $form_state, $old_form);
-    return $form;
-  }
-
-}
-
-/**
- * Long annotation lines over 80 characters are ok.
- *
- * @WebformHandler(
- *   id = "email",
- *   label = @Translation("Email"),
- *   category = @Translation("Notification"),
- *   description = @Translation("Sends a webform submission via an email."),
- *   cardinality = \Drupal\webform\WebformHandlerInterface::CARDINALITY_UNLIMITED,
- *   results = \Drupal\webform\WebformHandlerInterface::RESULTS_PROCESSED,
- * )
- */
-class EmailWebformHandler extends WebformHandlerBase implements WebformHandlerMessageInterface {
-
-}
-
-/**
- * Deprecation check.
- *
- * @deprecated in drupal:8.7.0 and is removed from drupal:9.0.0.
- * Switch off your television set and go and do something less boring instead.
- * @see http://www.drupal.org/node/123
- */
-class Test4 {
-
-  /**
-   * Check trigger_error format.
-   */
-  public function testDeprecation() {
-    @trigger_error('Function testDeprecation() is deprecated in drupal:8.5.0 and is removed from drupal:9.0.0. Why Dont You. See http://www.drupal.org/node/123', E_USER_DEPRECATED);
-  }
-
-}
-
-/**
- * Foo.
- */
-interface Test5Interface {
-
-  /**
-   * Orders the result set by a given field.
-   *
-   * If called multiple times, the query will order by each specified field in
-   * the order this method is called.
-   *
-   * If the query uses DISTINCT or GROUP BY conditions, fields or expressions
-   * that are used for the order must be selected to be compatible with some
-   * databases like PostgreSQL. The PostgreSQL driver can handle simple cases
-   * automatically but it is suggested to explicitly specify them. Additionally,
-   * when ordering on an alias, the alias must be added before orderBy() is
-   * called.
-   *
-   * @param string $field
-   *   The field on which to order. The field is escaped for security so only
-   *   valid field and alias names are possible. To order by an expression, add
-   *   the expression with addExpression() first and then use the alias to order
-   *   on.
-   *
-   *   Example:
-   *   @code
-   *   $query->addExpression('SUBSTRING(thread, 1, (LENGTH(thread) - 1))', 'order_field');
-   *   $query->orderBy('order_field', 'ASC');
-   *   @endcode
-   * @param string $direction
-   *   The direction to sort. Legal values are "ASC" and "DESC". Any other value
-   *   will be converted to "ASC".
-   *
-   * @return \Drupal\Core\Database\Query\SelectInterface
-   *   The called object.
-   */
-  public function orderBy($field, $direction = 'ASC');
-
-  /**
-   * Example with multiple code blocks in param docs.
-   *
-   * @param string $param1
-   *   Just some Example param.
-   * @param ...
-   *   Any additional arguments are passed on to the functions called by
-   *   self::submitForm(), including the unique form constructor function.
-   *   For example, the node_edit form requires that a node object be passed
-   *   in here when it is called. Arguments that need to be passed by reference
-   *   should not be included here, but rather placed directly in the
-   *   $form_state build info array so that the reference can be preserved. For
-   *   example, a form builder function with the following signature:
-   *   @code
-   *   function mymodule_form($form, FormStateInterface &$form_state, &$object) {
-   *   }
-   *   @endcode
-   *   would be called via self::submitForm() as follows:
-   *   @code
-   *   $form_state->setValues($my_form_values);
-   *   $form_state->addBuildInfo('args', [&$object]);
-   *   \Drupal::formBuilder()->submitForm('mymodule_form', $form_state);
-   *   @endcode
-   */
-  public function test1($param1);
-
-  /**
-   * This is an example of a doc block that is good.
-   *
-   * We want to show some example code:
-   * @code
-   *   if ($something) {
-   *     $x = $y;
-   *   }
-   * @endcode
-   * Some more example code:
-   * @code
-   *   if ($something) {
-   *     $x = $y;
-   *   }
-   * @endcode
-   * And one more piece of example code:
-   * @code
-   *   if ($something) {
-   *     $x = $y;
-   *   }
-   * @endcode
-   * Followed by some summary text.
-   */
-  public function test2();
-
-  /**
-   * This is good.
-   *
-   * @return string
-   *   Here is a comment, let's explain the return value with an example:
-   *   @code
-   *     if ($something) {
-   *       $x = $y;
-   *     }
-   *   @endcode
-   *   And then the comment goes on here. You want more code? Here you go:
-   *   @code
-   *     if ($something) {
-   *       $x = $y;
-   *     }
-   *   @endcode
-   *   And this is the end.
-   */
-  public function test3();
-
-}
-
-/**
- * Test PHP attributes.
- */
-class TestPhpAttributes {
-
-  /**
-   * Bar property.
-   */
-  #[NotBlank]
-  private bool $bar;
-
-  /**
-   * Tests method with PHP attribute and docblock.
-   */
-  #[\ReturnTypeWillChange]
-  public function attributes(): void {
-  }
-
-}
-
-/**
- * Enums must have a comment block.
- */
-enum MyWellNamedEnum: int {
-  case One = 1;
-  case Two = 2;
-}
-
 // Allow longer than 80 character array definitions on one line (up to 120).
 if (!in_array($form_object->getOperation(), ['add', 'edit', 'delete', 'clone', 'default'], TRUE)) {
   return;
 }
 if ($content_type_facet_field && in_array($entity_type, ['entity:node', 'entity:media'])) {
   $foo->setContentTypeFacetValues($item, $entity_type, $content_type_facet_field);
-}
-
-/**
- * Testing nested array definitions going over 80 characters.
- */
-class LongNestedArrayLine {
-
-  /**
-   * Test method.
-   */
-  public function foo() {
-    foreach ($x as $y) {
-      foreach ($a as $b) {
-        $form[$policyTypeKey]['directives'][$directiveName]['options']['flags_wrapper']['flags'] = [
-          '#type' => 'checkboxes',
-          '#parents' => [$policyTypeKey, 'directives', $directiveName, 'flags', 'humans'],
-          '#options' => [
-            'unsafe-inline' => "'unsafe-inline'",
-            'unsafe-eval' => "'unsafe-eval'",
-            'unsafe-hashes' => "'unsafe-hashes'",
-            'unsafe-allow-redirects' => "'unsafe-allow-redirects'",
-            'strict-dynamic' => "'strict-dynamic'",
-            'report-sample' => "'report-sample'",
-          ],
-          '#default_value' => $config->get($policyTypeKey . '.directives.' . $directiveName . '.flags') ?: [],
-        ];
-      }
-    }
-  }
-
 }
 
 // Allow array definition statements over 80 but under 120 characters.
@@ -1878,118 +1217,10 @@ $form['strings'] = [
 ];
 
 /**
- * Implements hook_cron().
- */
-#[Hook('cron')]
-class CronHook {
-
-  public function __construct(
-    private readonly EntityTypeManagerInterface $entityTypeManager,
-    private readonly StreamWrapperManagerInterface $streamWrapperManager,
-    private readonly ConfigFactoryInterface $configFactory,
-    private readonly FileUsageInterface $fileUsage,
-    private readonly TimeInterface $time,
-    #[Autowire('@logger.channel.file')]
-    private readonly LoggerInterface $logger,
-  ) {}
-
-}
-
-/**
- * Doc block is here and an ignore directive is ok.
- */
-// phpcs:ignore Drupal.NamingConventions.ValidClassName
-enum PUROSELY_WRONG_BUT_OK: int {
-  case One = 1;
-  case Two = 2;
-}
-
-/**
- * Fully qualified class name is allowed in PHP attributes for now.
- */
-#[\Drupal\action_link\Attribute\StateAction(
-  id: 'test_always',
-  label: new \Drupal\Core\StringTranslation\TranslatableMarkup('Test Always'),
-  description: new \Drupal\Core\StringTranslation\TranslatableMarkup('Test Always'),
-  directions: [
-    'change' => 'change',
-  ]
-)]
-class TestAlways extends StateActionBase {
-
-  /**
-   * Partial names are ok in attributes for now.
-   */
-  #[Assert\NotBlank]
-  private bool $bar;
-
-  /**
-   * Partially qualified names are ok in attributes for now.
-   */
-  #[CLI\Command(
-    name: 'example',
-    aliases: ['example-foo']
-  )]
-  #[CLI\Option(name: 'pretty_format', description: 'Display the count in pretty format.')]
-  public function test(array $options = ['pretty-format' => TRUE]): void {
-  }
-
-}
-
-/**
  * Doc block is here and an ignore directive is ok.
  */
 // phpcs:ignore Drupal.NamingConventions.ValidClassName
 function phpcs_ignore_comment() {
-
-}
-
-/**
- * Test class.
- */
-class TestPlugin {
-
-  /**
-   * Gets a fallback id for a missing plugin.
-   *
-   * This method should be implemented in extending classes that also implement
-   * FallbackPluginManagerInterface. It is called by
-   * PluginManagerBase::handlePluginNotFound on the abstract class, and
-   * therefore should be defined as well on the abstract class to prevent static
-   * analysis errors.
-   *
-   * @param string $plugin_id
-   *   The ID of the missing requested plugin.
-   * @param array $configuration
-   *   An array of configuration relevant to the plugin instance.
-   *
-   * phpcs:ignore Drupal.Commenting.FunctionComment.InvalidNoReturn
-   * @return string
-   *   The id of an existing plugin to use when the plugin does not exist.
-   *
-   * @throws \BadMethodCallException
-   *   If the method is not implemented in the concrete plugin manager class.
-   */
-  protected function getFallbackPluginId($plugin_id, array $configuration = []) {
-    throw new \BadMethodCallException(static::class . '::getFallbackPluginId() not implemented.');
-  }
-
-}
-
-/**
- * Provides a collection of condition plugins.
- */
-class ConditionPluginCollection extends DefaultLazyPluginCollection {
-
-  /**
-   * {@inheritdoc}
-   *
-   * phpcs:ignore Drupal.Commenting.FunctionComment.MissingReturnComment
-   * @return \Drupal\Core\Condition\ConditionInterface
-   */
-  public function &get($instance_id) {
-    return 'x';
-  }
 
 }
 
@@ -2011,33 +1242,6 @@ public function ignore_phpstan_comment_with_attribute_before() {
 }
 
 /**
- * Test PHPCS ignore comments between param docs.
- */
-interface BreadcrumbBuilderInterface {
-
-  /**
-   * Whether this breadcrumb builder should be used to build the breadcrumb.
-   *
-   * @param \Drupal\Core\Routing\RouteMatchInterface $route_match
-   *   The current route match.
-   * phpcs:ignore Drupal.Commenting.FunctionComment.ParamNameNoMatch
-   * @param \Drupal\Core\Cache\CacheableMetadata $cacheable_metadata
-   *   The cacheable metadata to add to if your check varies by or depends
-   *   on something. Anything you specify here does not have to be repeated in
-   *   the build() method as it will be merged in automatically.
-   *
-   * @return bool
-   *   TRUE if this builder should be used or FALSE to let other builders
-   *   decide.
-   *
-   * @todo Uncomment new method parameters before drupal:12.0.0, see
-   *   https://www.drupal.org/project/drupal/issues/3459277.
-   */
-  public function applies(RouteMatchInterface $route_match /* , CacheableMetadata $cacheable_metadata */);
-
-}
-
-/**
  * Test that nested array types are ok.
  *
  * @param array<array<scalar|null>|object|scalar|null> $param
@@ -2054,23 +1258,3 @@ function pdo_weird_return_type($param) {
  * Comments are allowed to end in 3 dots...
  */
 function comment_test_dots() {}
-
-/**
- * Executes the page caching before the main kernel takes over the request.
- */
-class PageCache implements HttpKernelInterface {
-
-  /**
-   * The wrapped HTTP kernel.
-   */
-  protected \Closure $httpKernel;
-
-  /**
-   * The entity for this result.
-   *
-   * @var \Drupal\Core\Entity\EntityInterface
-   */
-  // phpcs:ignore Drupal.NamingConventions.ValidVariableName.LowerCamelName,PSR2.Classes.PropertyDeclaration.Underscore
-  public $_entity = NULL;
-
-}
